Initiali XML deserialization implementation

+++ lib/xommelier/core_ext/numeric.rb
@@ -1,4 +1,4 @@
-class Integer
+class Numeric
   def self.from_xommelier(value)
     return nil if value.blank?
     begin
